Add playnext command (Close #39). Updating audio.js functions to use promises, modify spotifyHandler and play functions to return song arrays instead of actually adding the track (The adding is done by the actual command. Implement better collectors (Close #16, Close  #19).

+++ commands/volume.js
@@ -12,7 +12,7 @@ module.exports = {
     guildOnly: true,
     execute(message, args)
     {
-        if (parseInt(args) != args) return message.channel.send(`Please provide a number!`);
+        if (parseFloat(args) != args) return message.channel.send(`Please provide a number!`);
 
         const currentQueue = am.getQueue(message);
